Déplace le code relatif aux suggestions dans un fichier dédié
