+++ zds/tutorialv2/forms.py
@@ -12,7 +12,7 @@ from zds.utils.forms import CommonLayoutEditor, CommonLayoutVersionEditor
 from zds.utils.models import SubCategory
 from zds.tutorialv2.models import TYPE_CHOICES
 from zds.tutorialv2.models.help_requests import HelpWriting
-from zds.tutorialv2.models.database import PublishableContent, ContentContributionRole, ContentSuggestion
+from zds.tutorialv2.models.database import PublishableContent, ContentContributionRole
 from django.utils.translation import gettext_lazy as _
 from zds.member.models import Profile
 from zds.utils.forms import IncludeEasyMDE
@@ -1186,53 +1186,6 @@ class ContentCompareStatsURLForm(forms.Form):
             raise forms.ValidationError(_("Vous devez choisir des URL a comparer"))
         if len(urls) < 2:
             raise forms.ValidationError(_("Il faut au minimum 2 urls à comparer"))
-
-
-class SearchSuggestionForm(forms.Form):
-    suggestion_pk = forms.CharField(
-        label="Contenu à suggérer",
-        required=False,
-        widget=forms.TextInput(),
-    )
-    excluded_pk = forms.CharField(required=False, widget=forms.HiddenInput(attrs={"class": "excluded_field"}))
-
-    def __init__(self, content, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        self.fields["suggestion_pk"].widget.attrs.update(
-            {
-                "data-autocomplete": '{"type": "multiple_checkbox",'
-                '"limit": 10,'
-                '"fieldname": "title",'
-                '"url": "' + reverse("search:suggestion") + '?q=%s&excluded=%e"}',
-                "placeholder": "Rechercher un contenu",
-            }
-        )
-
-        self.helper = FormHelper()
-        self.helper.form_action = reverse("content:add-suggestion", kwargs={"pk": content.pk})
-        self.helper.form_class = "modal modal-large"
-        self.helper.form_id = "add-suggestion"
-        self.helper.form_method = "post"
-
-        self.helper.layout = Layout(
-            Field("suggestion_pk"), Field("excluded_pk"), StrictButton(_("Ajouter"), type="submit")
-        )
-
-
-class RemoveSuggestionForm(forms.Form):
-    pk_suggestion = forms.IntegerField(
-        label=_("Suggestion"),
-        required=True,
-        error_messages={"does_not_exist": _("La suggestion sélectionnée n'existe pas.")},
-    )
-
-    def clean_pk_suggestion(self):
-        pk_suggestion = self.cleaned_data.get("pk_suggestion")
-        suggestion = ContentSuggestion.objects.filter(id=pk_suggestion).first()
-        if suggestion is None:
-            self.add_error("pk_suggestion", self.fields["pk_suggestion"].error_messages["does_not_exist"])
-        return pk_suggestion
 
 
 class ToggleHelpForm(forms.Form):
