+++ zds/tutorialv2/tests/tests_views/tests_removesuggestion.py
@@ -7,7 +7,7 @@ from django.utils.html import escape
 
 from zds.member.tests.factories import ProfileFactory, StaffProfileFactory
 from zds.tutorialv2.tests.factories import PublishableContentFactory
-from zds.tutorialv2.forms import RemoveSuggestionForm
+from zds.tutorialv2.views.suggestions import RemoveSuggestionForm
 from zds.tutorialv2.models.database import ContentSuggestion
 from zds.tutorialv2.tests import TutorialTestMixin, override_for_contents
