+++ zds/tutorialv2/views/display/content.py
@@ -24,9 +24,9 @@ from zds.tutorialv2.forms import (
     PickOpinionForm,
     UnpickOpinionForm,
     PromoteOpinionToArticleForm,
-    SearchSuggestionForm,
     ContributionForm,
 )
+from zds.tutorialv2.views.suggestions import SearchSuggestionForm
 from zds.tutorialv2.views.licence import EditContentLicenseForm
 from zds.tutorialv2.views.tags import EditTagsForm
 from zds.tutorialv2.mixins import SingleContentDetailViewMixin, SingleOnlineContentDetailViewMixin
